assortment_chatbot/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "assortment-chatbot"
version = "0.2.0"
description = "A web-based chatbot application for data exploration and analysis powered by PydanticAI"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Data Chat Team" }]
keywords = ["streamlit", "dashboard", "visualization", "chat", "LLM", "AI"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
]

dependencies = [
    "streamlit>=1.25.0",
    "pygwalker>=0.4.9",
    "duckdb>=0.8.0",
    "pydantic>=2.1.0",
    "pydantic-settings>=2.0.0",
    "pymongo>=4.4.0",
    "snowflake-connector-python>=3.0.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "cryptography>=41.0.0",
    "xlsxwriter>=3.1.0",
    "openpyxl>=3.1.0",
    "fastparquet>=2023.10.0",
    "streamlit-option-menu>=0.4.0",
    "matplotlib>=3.10.0",
    "plotly>=6.0.0",
    "openai>=1.75.0",
    "pydantic-ai",
    "azure-keyvault-secrets>=4.9.0",
    "azure-identity>=1.21.0",
    "loguru>=0.7.0",
]

[project.urls]
"Source Code" = "https://github.com/yourusername/assortment-chatbot"
"Issue Tracker" = "https://github.com/yourusername/assortment-chatbot/issues"
"Documentation" = "https://yourusername.github.io/assortment-chatbot"

[project.scripts]
assortment-chatbot = "assortment_chatbot.app:main"

[tool.setuptools]
packages = ["assortment_chatbot"]

[tool.setuptools.package-data]
"assortment_chatbot" = ["*.json", "*.yaml", "*.css"]

[tool.ruff]
target-version = "py310"
line-length = 100
fix = true
unsafe-fixes = true
extend-select = ["E", "F", "I", "N", "B", "UP", "C", "A", "T", "Q"]
extend-ignore = ["E203", "E501"]

[tool.ruff.lint.isort]
known-first-party = ["assortment_chatbot"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["streamlit.*", "pywalker.*", "duckdb.*", "snowflake.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
